# code/sonara_back_end/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Sonara API",
    "description": "\nAPI do projeto **Sonara** — plataforma de conexão entre artistas e organizadores de eventos.\n\n## Autenticação\nAlgumas rotas exigem **Bearer Token** (JWT). Faça login em `POST /v1/sonara/usuario/login` e use o token retornado no header:\n`Authorization: Bearer <token>`\n\n## Fluxo principal\n1. Cadastre um usuário (artista ou organizador)\n2. Faça login para obter o token\n3. Crie eventos (organizador) ou candidate-se a eventos (artista)\n4. Gerencie candidaturas, contra-propostas e convites\n        ",
    "version": "1.0.10",
    "contact": {
      "name": "Davi de Almeida Santos"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Insira: Bearer <token>"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/sonara/artista/": {
      "get": {
        "tags": [
          "Artista"
        ],
        "summary": "Listar todos os artistas",
        "description": "",
        "responses": {
          "200": {
            "description": "Lista retornada com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Nenhum artista encontrado"
          }
        }
      },
      "post": {
        "tags": [
          "Artista"
        ],
        "summary": "Cadastrar artista",
        "description": "",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Artista"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Artista criado com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Campos inválidos"
          },
          "415": {
            "description": "Content-Type deve ser application/json"
          }
        }
      }
    },
    "/v1/sonara/artista/{id}": {
      "get": {
        "tags": [
          "Artista"
        ],
        "summary": "Buscar artista por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do artista"
          }
        ],
        "responses": {
          "200": {
            "description": "Artista encontrado"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "ID inválido"
          },
          "404": {
            "description": "Artista não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Artista"
        ],
        "summary": "Atualizar artista",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do artista"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Artista"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artista atualizado"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Artista não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Artista"
        ],
        "summary": "Excluir artista",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do artista"
          }
        ],
        "responses": {
          "200": {
            "description": "Artista excluído"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Artista não encontrado"
          }
        }
      }
    },
...
    "/v1/sonara/evento/": {
      "get": {
        "tags": [
          "Evento"
        ],
        "summary": "Listar todos os eventos",
        "description": "Retorna todos os eventos cadastrados com fotos, endereço e organizador.",
        "responses": {
          "200": {
            "description": "Lista de eventos retornada com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Nenhum evento encontrado"
          }
        }
      },
      "post": {
        "tags": [
          "Evento"
        ],
        "summary": "Criar evento",
        "description": "Cria evento com endereço e vínculo ao organizador.",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventoCadastro"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Evento criado com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Campos obrigatórios ausentes ou inválidos"
          },
          "415": {
            "description": "Content-Type deve ser application/json"
          }
        }
      }
    },
    "/v1/sonara/evento/{id}": {
      "get": {
        "tags": [
          "Evento"
        ],
        "summary": "Buscar evento por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do evento"
          }
        ],
        "responses": {
          "200": {
            "description": "Evento encontrado"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Evento"
        ],
        "summary": "Atualizar evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do evento"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventoCadastro"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento atualizado com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Evento"
        ],
        "summary": "Excluir evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do evento"
          }
        ],
        "responses": {
          "200": {
            "description": "Evento excluído com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        }
      }
    },
...
    "/v1/sonara/avaliacaoArtista/": {
      "get": {
        "tags": [
          "Avaliacao Artista"
        ],
        "summary": "Listar avaliações de artistas",
        "description": "",
        "responses": {
          "200": {
            "description": "Lista retornada"
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "post": {
        "tags": [
          "Avaliacao Artista"
        ],
        "summary": "Criar avaliação de artista",
        "description": "",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AvaliacaoArtista"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Avaliação criada"
          },
          "204": {
            "description": "No Content"
          },
          "415": {
            "description": "Content-Type deve ser application/json"
          }
        }
      }
    },
    "/v1/sonara/avaliacaoArtista/{id}": {
      "get": {
        "tags": [
          "Avaliacao Artista"
        ],
        "summary": "Buscar avaliação de artista por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Encontrado"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Avaliacao Artista"
        ],
        "summary": "Atualizar avaliação de artista",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AvaliacaoArtista"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Avaliação atualizada"
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Avaliacao Artista"
        ],
        "summary": "Excluir avaliação de artista",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Avaliação excluída"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/v1/sonara/avaliacaoEvento/": {
      "get": {
        "tags": [
          "Avaliacao Evento"
        ],
        "summary": "Listar avaliações de eventos",
        "description": "",
        "responses": {
          "200": {
            "description": "Lista retornada"
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "post": {
        "tags": [
          "Avaliacao Evento"
        ],
        "summary": "Criar avaliação de evento",
        "description": "",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AvaliacaoEvento"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Avaliação criada"
          },
          "204": {
            "description": "No Content"
          },
          "415": {
            "description": "Content-Type deve ser application/json"
          }
        }
      }
    },
    "/v1/sonara/avaliacaoEvento/{id}": {
      "get": {
        "tags": [
          "Avaliacao Evento"
        ],
        "summary": "Buscar avaliação de evento por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Avaliação encontrada"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Não encontrada"
          }
        }
      },
      "put": {
        "tags": [
          "Avaliacao Evento"
        ],
        "summary": "Atualizar avaliação de evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AvaliacaoEvento"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Avaliação atualizada"
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Avaliacao Evento"
        ],
        "summary": "Excluir avaliação de evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Avaliação excluída"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "EventoCadastro": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "example": "Festival Sonara 2026"
        },
        "descricao": {
          "type": "string",
          "example": "Festival de música ao ar livre"
        },
        "local": {
          "type": "string",
          "example": "Parque Ibirapuera"
        },
        "data": {
          "type": "string",
          "example": "2026-08-15"
        },
        "hora_inicio": {
          "type": "string",
          "example": "18:00"
        },
        "hora_fim": {
          "type": "string",
          "example": "23:00"
        },
        "organizador_id": {
          "type": "number",
          "example": 1
        },
        "cep": {
          "type": "string",
          "example": "04094-050"
        },
        "cidade": {
          "type": "string",
          "example": "São Paulo"
        },
        "estado": {
          "type": "string",
          "example": "SP"
        },
        "logradouro": {
          "type": "string",
          "example": "Av. Pedro Álvares Cabral"
        },
        "numero": {
          "type": "string",
          "example": "S/N"
        },
        "bairro": {
          "type": "string",
          "example": "Vila Mariana"
        },
        "complemento": {
          "type": "string",
          "example": "Portão 3"
        }
      },
      "required": [
        "nome",
        "descricao",
        "local",
        "data",
        "hora_inicio",
        "hora_fim",
        "organizador_id",
        "cep",
        "cidade",
        "estado",
        "logradouro",
        "numero",
        "bairro"
      ]
    },
    "AvaliacaoArtista": {
      "type": "object",
      "properties": {
        "numero_estrelas": {
          "type": "number",
          "example": 5
        },
        "usuario_id": {
          "type": "number",
          "example": 1
        },
        "artista_id": {
          "type": "number",
          "example": 1
        },
        "data_avaliacao": {
          "type": "string",
          "example": "2026-06-01 20:00:00"
        }
      },
      "required": [
        "numero_estrelas",
        "usuario_id",
        "artista_id",
        "data_avaliacao"
      ]
    },
    "AvaliacaoEvento": {
      "type": "object",
      "properties": {
        "numero_estrelas": {
          "type": "number",
          "example": 4
        },
        "usuario_id": {
          "type": "number",
          "example": 1
        },
        "evento_id": {
          "type": "number",
          "example": 1
        },
        "data_avaliacao": {
          "type": "string",
          "example": "2026-06-01 22:00:00"
        }
      },
      "required": [
        "numero_estrelas",
        "usuario_id",
        "evento_id",
        "data_avaliacao"
      ]
    },
...
    "Artista": {
      "type": "object",
      "properties": {
        "nome_artistico": {
          "type": "string",
          "example": "DJ João"
        },
        "usuario_id": {
          "type": "number",
          "example": 1
        },
        "descricao": {
          "type": "string",
          "example": "DJ com 5 anos de experiência"
        }
      },
      "required": [
        "nome_artistico",
        "usuario_id",
        "descricao"
      ]
    },
...
  }
}